requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.11.0
starlette>=0.46.0

# Environment variables management
python-dotenv>=1.0.0

# HTTP client for API calls
httpx>=0.28.0

# LLM and RAG components
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
faiss-cpu>=1.7.4

# LLM integrations
openai>=1.3.5
langchain-groq>=0.1.0

# Reinforcement Learning (lightweight)
stable-baselines3>=2.2.1
gymnasium>=1.0.0

# Data processing
pandas>=2.1.3
numpy>=1.26.2

# WebSocket support for real-time data
websockets>=15.0.0

# JSON processing
json5>=0.9.14

# Autogen for multi-agent system
pyautogen>=0.2.3

# Testing
pytest>=7.4.3

# Async utilities
asyncio>=3.4.3
aiohttp>=3.9.0

python-telegram-bot==22.1